added test to suite
