+++ server/package.json
@@ -8,7 +8,9 @@
   },
   "scripts": {
     "start": "node index.js",
-    "test": "jest --passWithNoTests"
+    "jest-test": "jest --passWithNoTests",
+    "extension-test": "npx mocha extension_test.js",
+    "test": "npm-run-all jest-test extension-test"
   },
   "repository": {
     "type": "git",
@@ -24,9 +26,11 @@
     "chai": "^4.2.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "mocha": "^7.1.2",
     "mongodb": "^3.5.3",
     "mongoose": "^5.9.6",
     "multer": "^1.4.2",
+    "npm-run-all": "^4.1.5",
     "sinon": "^8.1.1",
     "supertest": "^4.0.2"
   },
